Test and Fix (round 3) - fix issue of qsfs not synchornizing the dir tree when editing through console while qsfs is running To reproduce the issue, 1) cd to root and ls; 2) remove a folder through web console; 3) ls again. The removed folder is still listed in qsfs.
Root Cause: The previous logic of updating dir is based on if the dir mtime is modified, and for root there is no mtime, root't mtime is set when connect to it at first time. As in object storage changing child nodes of a folder (such as root or a dir node) will not update the folder mtime, so such logic will fail.

Resolution: Always update dir synchronizley in opendir, and only in opendir. so, in getattr set updateifdir to false. QSClient::Stat only update node itself with modified meta, QSClient::ListDirectory always update dir and move the logic of deciding when to updatedir in upper level.